pyproject.toml
[project]
name = "hft-simulation"
version = "0.1.0"
description = "HFT Simulation"
readme = "README.md"
authors = [{ name = "mlapid-ya", email = "m.lapid@yandex.ru" }]
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "websockets>=14.2",
]

[project.scripts]
hft-simulation = "hft_simulation:main"
exchange-connector = "hft_simulation.exchange_connector:main"
processing-engine = "hft_simulation.processing_engine:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["aiosqlite>=0.21.0", "pytest>=8.3.4", "pytest-asyncio>=0.25.3"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
